Add crop preview functionality
* CZPhotoPickerController now accepts a CGSize to set the desired
size of the crop overlay view. Images are centered beneath the crop
overlay view to give you a sense of how the image will be cropped
when it is displayed in an image view of the same size. The edited
image that is returned is cropped such that the aspect ratio matches
the aspect ratio of the provided size.