Make preview overview's mask be gradient
The reason for using gradient is that we are setting up camera overlay
view (`cameraOverlayView`) to our custom view, but we still want to
keep using standard camera controls. The documentation actually says:

> The image picker layers your custom overlay view on top of the other
> image picker views and positions it relative to the screen
> coordinates. If you have the default camera controls set to be
> visible, incorporate transparency into your view, or position it to
> avoid obscuring the underlying content.

But sadly there is no way to get the frame of camera controls views
without hacks. As a result CZCropPreviewOverlayView was always over all
other controls and as it had background with alpha 0.6, camera controls
looked dimmed. Plus we don't seem to have much control over
cameraOverlayView's position and as a result in standard preview mode it
was slightly in a wrong position.

The solution here is to make our mask to fade out towards the top and
bottom edge of the screen so the standard camera controls aren't
"dimmed" and user can still notice the cropping.